+++ Assets/Scripts/Monster/MonsterData.cs
@@ -33,6 +33,6 @@ public class MonsterData : ScriptableObject
     [SerializeField] float qSkillStiffTime;
     public float QSkillStiffTime { get { return qSkillStiffTime; } }
 
-    [SerializeField] float exp;
-    public float EXP { get { return exp; } }
+    [SerializeField] int exp;
+    public int EXP { get { return exp; } }
 }
